Rebuilding a BOM entry to the output file
Initially, the implementation of the program meant checking whether there is the BOM in the output file. 
Then we remembered it, and after deleting all the data in the file, we wrote the BOM to it, if it exists.
Now the BOM is recorded based only on the input file.