Modify the read command and bundles
Modifying the read command and arguments as follows:
1) The characters '<', ' > ' or '>> ' are treated as separators.
2) Input and output Redirection defined by the characters‘<', '>' or '>>'.
3) Implementation a pipeline with an unlimited number of commands using '|'.
4) Recognizing the '&' character (if it is encountered outside of quotation marks) as a separator character and implementing command execution in the background.
The command runs in the background with the '& ' character as the last word. 
If the '&' character is not found at the end, the program returns an error message. 
When a command that was running in the background has completed, a message about its completion and a completion code are displayed.

Next, bundles are implemented:
1) a bunch of ';' (sequential execution of commands).
2) a bunch of '||' (the first command is executed first, and if it fails, the second command is executed).
3) a bunch of '&&' (the first command is executed first, and if it is successful, the second command is executed).